update sync video flow

+++ app/src/main/java/com/celiluysal/watchtogetherapp/ui/room/playlist_dialog/VideoRecyclerViewAdapter.kt
@@ -1,7 +1,8 @@
-package com.celiluysal.watchtogetherapp.ui.room.playlist
+package com.celiluysal.watchtogetherapp.ui.room.playlist_dialog
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
+import android.widget.ImageView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.celiluysal.watchtogetherapp.databinding.ItemVideoCardBinding
@@ -9,28 +10,33 @@ import com.celiluysal.watchtogetherapp.models.WTVideo
 
 class VideoRecyclerViewAdapter(
     private val playList: MutableList<WTVideo>,
+    private val userIsOwner: Boolean,
     var clickListener: onVideoItemClickListener
 ) : RecyclerView.Adapter<VideoRecyclerViewAdapter.VideoViewHolder>() {
 
-    class VideoViewHolder(val binding: ItemVideoCardBinding) :
+    inner class VideoViewHolder(val binding: ItemVideoCardBinding) :
         RecyclerView.ViewHolder(binding.root) {
 
         fun bind(wtVideo: WTVideo, action: onVideoItemClickListener) {
 
             binding.textViewVideoTitle.text = wtVideo.title
-//            binding.textViewVideoTitle.isSelected = true
             binding.textViewChannel.text = wtVideo.channel
 
             Glide.with(binding.root).load(wtVideo.thumbnail)
                 .into(binding.imageViewThumbnail)
 
-            binding.imageViewDelete.setOnClickListener {
-                action.onDeleteClick(wtVideo, adapterPosition)
+            if (userIsOwner) {
+                binding.imageViewDelete.setOnClickListener {
+                    action.onDeleteClick(wtVideo, adapterPosition)
+                }
+
+                binding.cardViewVideo.setOnClickListener{
+                    action.onVideoItemClick(wtVideo, adapterPosition)
+                }
+            } else {
+                binding.imageViewDelete.visibility = ImageView.GONE
             }
 
-            binding.cardViewVideo.setOnClickListener{
-                action.onVideoItemClick(wtVideo, adapterPosition)
-            }
         }
     }
